add sorting and filter

+++ DeviceRepository/Common/Page/PageInfo.cs
@@ -2,14 +2,14 @@ namespace DeviceRepository.Common.Page;
 
 public class PageInfo
 {
-    public int Size { get; set; }
-    public int Page { get; set; }
+    public ushort Size { get; set; }
+    public ushort Page { get; set; }
 
     public PageInfo() : this(20, 1)
     {
     }
 
-    public PageInfo(int size, int page)
+    public PageInfo(ushort size, ushort page)
     {
         Size = size;
         Page = page;
